app/requirements.txt
pygithub==1.59.0
gitpython==3.1.31
python-dotenv==1.0.0